+++ packages/Bioreactorworkflowdemo/src/package-test.ts
@@ -1,5 +1,6 @@
 import { runTests, tests, TestContext , initAutoTests as initTests } from '@datagrok-libraries/test/src/test';
 import * as DG from 'datagrok-api/dg';
+import './tests/bioreactor-model-tests';
 
 export let _package = new DG.Package();
 export { tests };